# sql: avoid string formatting in reportSessionDataChanges when not necessary